--- src/main/java/CaviarException.java
@@ -1,5 +1,0 @@
-public class CaviarException extends Exception {
-    public CaviarException(String message) {
-        super("roe..!! " + message);
-    }
-}
