+++ src/main/java/caviar/task/Task.java
@@ -1,4 +1,13 @@
-import java.io.IOException;
+package caviar.task;
+
+import caviar.storage.Storage;
+import caviar.task.Deadline;
+import caviar.task.Event;
+import caviar.task.Todo;
+import caviar.command.TaskList;
+import caviar.exception.CaviarException;
+
+import caviar.exception.CaviarException;
 
 public class Task {
     protected String description;
