+++ src/main/java/caviar/task/Todo.java
@@ -1,0 +1,19 @@
+package caviar.task;
+
+import caviar.storage.Storage;
+import caviar.task.Deadline;
+import caviar.task.Task;
+import caviar.task.Event;
+import caviar.command.TaskList;
+import caviar.exception.CaviarException;
+
+public class Todo extends Task {
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
