+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm Caviar. Roe!
+Hello! I'm caviar.Caviar. Roe!
 What can I do for you?
 ______________________
